z1cars_poster_saas_v3_final/requirements.txt
flask
pillow
opencv-python-headless
requests
